+++ test/unit/core/component/brushing.spec.js
@@ -103,7 +103,7 @@ describe('Brushing', () => {
 
       resolveTapEvent({ e: eventMock, t: trigger, config });
 
-      expect(brushContext.toggleValues).to.not.have.been.called;
+      expect(brushContext.toggleValues).to.have.been.calledWith([]);
     });
 
     it('should default to "self" if no data context is configured', () => {